+++ src/main/java/pairmatching/domain/program/Course.java
@@ -1,6 +1,10 @@
-package pairmatching;
+package pairmatching.domain.program;
 
+import pairmatching.ErrorConstants;
+
+import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.List;
 
 public enum Course {
     BACKEND("백엔드"),
@@ -18,5 +22,12 @@ public enum Course {
                 .findAny()
                 .orElseThrow(() ->
                         new IllegalArgumentException(ErrorConstants.ERROR_PREFIX + "정확한 코스 이름을 입력해주세요."));
+    }
+
+    public static List<String> getCourseNames() {
+        List<String> courseNames = new ArrayList<>();
+        Arrays.stream(values())
+                .forEach(course -> courseNames.add(course.name));
+        return courseNames;
     }
 }
